services/api/requirements.txt
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.0
sqlalchemy>=2.0
asyncpg>=0.29
alembic>=1.13
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
redis>=5.0
opentelemetry-api>=1.23
opentelemetry-sdk>=1.23
opentelemetry-instrumentation-fastapi>=0.44b
mlflow>=2.12
httpx>=0.27
